backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.9.2
